pyproject.toml
[project]
name = "ai-platform"
version = "0.1.0"
description = "Production-grade AI backend platform"
authors = [
    {name = "salmanfarooq1",email = "salmanfarooq673@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "aiohttp (>=3.13.3,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
